+++ src/modules/categories/categories.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { CategoriesService } from './services/categories.service';
+import { CategoriesController } from './categories.controller';
+
+@Module({
+  controllers: [CategoriesController],
+  providers: [CategoriesService],
+})
+export class CategoriesModule {}