+++ src/modules/categories/index.ts
@@ -1,0 +1,6 @@
+export * from './categories.controller';
+export * from './categories.module';
+export * from './dto/create-category.dto';
+export * from './dto/update-category.dto';
+export * from './entities/category.entity';
+export * from './services/categories.service';