+++ src/modules/likes/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/create-like.dto';
+export * from './dto/update-like.dto';
+export * from './entities/like.entity';
+export * from './likes.controller';
+export * from './likes.module';
+export * from './services/likes.service';