+++ src/modules/posts/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/create-post.dto';
+export * from './dto/update-post.dto';
+export * from './entities/post.entity';
+export * from './posts.controller';
+export * from './posts.module';
+export * from './services/posts.service';