+++ src/modules/timelines/dto/update-timeline.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateTimelineDto } from './create-timeline.dto';
+
+export class UpdateTimelineDto extends PartialType(CreateTimelineDto) {}