+++ src/modules/timelines/entities/timeline.entity.ts
@@ -1,0 +1,4 @@
+import { Entity } from 'typeorm';
+
+@Entity('timelines')
+export class Timeline {}