+++ src/modules/timelines/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/create-timeline.dto';
+export * from './dto/update-timeline.dto';
+export * from './entities/timeline.entity';
+export * from './services/timelines.service';
+export * from './timelines.controller';
+export * from './timelines.module';