+++ src/modules/timelines/timelines.controller.ts
@@ -1,0 +1,45 @@
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+} from '@nestjs/common';
+import { TimelinesService } from './services/timelines.service';
+import { CreateTimelineDto } from './dto/create-timeline.dto';
+import { UpdateTimelineDto } from './dto/update-timeline.dto';
+
+@Controller('timelines')
+export class TimelinesController {
+  constructor(private readonly timelinesService: TimelinesService) {}
+
+  @Post()
+  create(@Body() createTimelineDto: CreateTimelineDto) {
+    return this.timelinesService.create(createTimelineDto);
+  }
+
+  @Get()
+  findAll() {
+    return this.timelinesService.findAll();
+  }
+
+  @Get(':id')
+  findOne(@Param('id') id: string) {
+    return this.timelinesService.findOne(+id);
+  }
+
+  @Patch(':id')
+  update(
+    @Param('id') id: string,
+    @Body() updateTimelineDto: UpdateTimelineDto,
+  ) {
+    return this.timelinesService.update(+id, updateTimelineDto);
+  }
+
+  @Delete(':id')
+  remove(@Param('id') id: string) {
+    return this.timelinesService.remove(+id);
+  }
+}