+++ src/modules/users/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/create-user.dto';
+export * from './dto/update-user.dto';
+export * from './entities/user.entity';
+export * from './services/users.service';
+export * from './users.controller';
+export * from './users.module';